# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WISR API",
    "description": "API documentation for the WISR application",
    "version": "V1"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Keycloak JWT authentication"
      }
    },
    "schemas": {
      "Age": {
        "type": "string",
        "enum": ["ADULT", "YOUNG", "UNKNOWN"],
        "description": "Age classification of the animal",
        "example": "ADULT"
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "description": "HTTP status code"
          },
          "error": {
            "type": "string",
            "description": "HTTP status text"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "description": "Error detail, replaced with a generic string for 5xx"
          }
        },
        "required": ["statusCode", "error", "message"],
        "additionalProperties": false,
        "description": "Standard error response"
      },
...
      "Incident": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "year": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "accidentDate": {
            "type": ["string", "null"]
          },
          "speciesId": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "speciesName": {
            "type": "string"
          },
          "speciesColor": {
            "type": "string"
          },
          "speciesGroupName": {
            "type": "string"
          },
          "serviceAreaId": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          },
          "serviceAreaName": {
            "type": ["string", "null"]
          },
          "contractAreaNumber": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          },
          "district": {
            "type": ["string", "null"]
          },
          "region": {
            "type": ["string", "null"]
          },
          "sex": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Sex"
              },
              {
                "type": "null"
              }
            ]
          },
          "timeOfKill": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TimeOfKill"
              },
              {
                "type": "null"
              }
            ]
          },
          "age": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Age"
              },
              {
                "type": "null"
              }
            ]
          },
          "quantity": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number",
                "minimum": -90,
                "maximum": 90
              },
              {
                "type": "null"
              }
            ]
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number",
                "minimum": -180,
                "maximum": 180
              },
              {
                "type": "null"
              }
            ]
          },
          "nearestTown": {
            "type": ["string", "null"]
          },
          "comments": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "year",
          "accidentDate",
          "speciesId",
          "speciesName",
          "speciesColor",
          "speciesGroupName",
          "serviceAreaId",
          "serviceAreaName",
          "contractAreaNumber",
          "district",
          "region",
          "sex",
          "timeOfKill",
          "age",
          "quantity",
          "latitude",
          "longitude",
          "nearestTown",
          "comments"
        ],
        "additionalProperties": false,
        "description": "A wildlife-vehicle collision record"
      },
      "IncidentCount": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "Incident records matching the filters"
          }
        },
        "required": ["total"],
        "additionalProperties": false,
        "description": "Count of matching incidents"
      },
      "IncidentFilters": {
        "type": "object",
        "properties": {
          "years": {
            "type": "array",
            "items": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "maximum": 9007199254740991
            }
          },
          "species": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Species"
            }
          },
          "serviceAreas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceArea"
            }
          },
          "sex": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Sex"
            }
          },
          "timeOfKill": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimeOfKill"
            }
          },
          "age": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Age"
            }
          },
          "dateRange": {
            "type": "object",
            "properties": {
              "min": {
                "anyOf": [
                  {
                    "type": "string",
                    "format": "date",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "max": {
                "anyOf": [
                  {
                    "type": "string",
                    "format": "date",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            },
            "required": ["min", "max"],
            "additionalProperties": false
          }
        },
        "required": [
          "years",
          "species",
          "serviceAreas",
          "sex",
          "timeOfKill",
          "age",
          "dateRange"
        ],
        "additionalProperties": false,
        "description": "Available filter values across the incident dataset"
      },
      "Incidents": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Incident"
            },
            "description": "Result items"
          },
          "total": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "Total matching records (before pagination)"
          },
          "limit": {
            "description": "Requested page size (omitted when unbounded)",
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "offset": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "description": "Number of records skipped"
          }
        },
        "required": ["data", "total", "offset"],
        "additionalProperties": false,
        "description": "Paginated incident records"
      },
...
      "ServiceArea": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "name": {
            "type": "string"
          },
          "contractAreaNumber": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "district": {
            "type": "string"
          },
          "region": {
            "type": "string"
          }
        },
        "required": ["id", "name", "contractAreaNumber", "district", "region"],
        "additionalProperties": false,
        "description": "Highway maintenance service area"
      },
...
      "Sex": {
        "type": "string",
        "enum": ["MALE", "FEMALE", "UNKNOWN"],
        "description": "Biological sex of the animal",
        "example": "MALE"
      },
      "Species": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "groupName": {
            "type": "string"
          }
        },
        "required": ["id", "name", "color", "groupName"],
        "additionalProperties": false,
        "description": "Wildlife species with display color and species group"
      },
      "TimeOfKill": {
        "type": "string",
        "enum": ["DAY", "DAWN", "DUSK", "DARK", "UNKNOWN"],
        "description": "Approximate time of day the incident occurred",
        "example": "DAY"
      }
    }
  },
  "paths": {
...
    "/v1/incidents/": {
      "get": {
        "operationId": "getIncidents",
        "summary": "Query wildlife-vehicle incidents",
        "tags": ["Incidents"],
        "description": "Returns wildlife-vehicle collision incidents with optional filtering by species, service area, date range, spatial geometry, and more.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated list of years (e.g. 2020,2021), required, max 10",
              "example": "2020,2021"
            },
            "in": "query",
            "name": "year",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated species IDs from GET /v1/incidents/filters",
              "example": "10,18"
            },
            "in": "query",
            "name": "species"
          },
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated service area IDs from GET /v1/incidents/filters",
              "example": "16,17"
            },
            "in": "query",
            "name": "serviceArea"
          },
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated values: MALE, FEMALE, UNKNOWN",
              "example": "MALE,FEMALE,UNKNOWN"
            },
            "in": "query",
            "name": "sex"
          },
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated values: DAY, DAWN, DUSK, DARK, UNKNOWN",
              "example": "DAY,DAWN,DUSK,DARK,UNKNOWN"
            },
            "in": "query",
            "name": "timeOfKill"
          },
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated values: ADULT, YOUNG, UNKNOWN",
              "example": "ADULT,YOUNG,UNKNOWN"
            },
            "in": "query",
            "name": "age"
          },
          {
            "schema": {
              "description": "Start date filter (inclusive, YYYY-MM-DD)",
              "example": "2021-01-01",
              "type": "string",
              "format": "date",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
            },
            "in": "query",
            "name": "startDate",
            "description": "Start date filter (inclusive, YYYY-MM-DD)"
          },
          {
            "schema": {
              "description": "End date filter (inclusive, YYYY-MM-DD)",
              "example": "2021-12-31",
              "type": "string",
              "format": "date",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
            },
            "in": "query",
            "name": "endDate",
            "description": "End date filter (inclusive, YYYY-MM-DD)"
          },
          {
            "schema": {
              "type": "string",
              "description": "GeoJSON geometry string for spatial filtering (e.g. Polygon)",
              "example": "{\"type\":\"Polygon\",\"coordinates\":[[[-123.2,49.2],[-123.0,49.2],[-123.0,49.3],[-123.2,49.3],[-123.2,49.2]]]}"
            },
            "in": "query",
            "name": "geometry"
          },
          {
            "schema": {
              "description": "Route corridor filter: start point longitude",
              "type": "number",
              "minimum": -180,
              "maximum": 180
            },
            "in": "query",
            "name": "routeStartLng",
            "description": "Route corridor filter: start point longitude"
          },
          {
            "schema": {
              "description": "Route corridor filter: start point latitude",
              "type": "number",
              "minimum": -90,
              "maximum": 90
            },
            "in": "query",
            "name": "routeStartLat",
            "description": "Route corridor filter: start point latitude"
          },
          {
            "schema": {
              "description": "Route corridor filter: end point longitude",
              "type": "number",
              "minimum": -180,
              "maximum": 180
            },
            "in": "query",
            "name": "routeEndLng",
            "description": "Route corridor filter: end point longitude"
          },
          {
            "schema": {
              "description": "Route corridor filter: end point latitude",
              "type": "number",
              "minimum": -90,
              "maximum": 90
            },
            "in": "query",
            "name": "routeEndLat",
            "description": "Route corridor filter: end point latitude"
          },
          {
            "schema": {
              "default": 250,
              "description": "Route corridor filter: corridor width in metres",
              "type": "number",
              "exclusiveMinimum": 0,
              "maximum": 20000
            },
            "in": "query",
            "name": "routeCorridorM",
            "description": "Route corridor filter: corridor width in metres"
          },
          {
            "schema": {
              "description": "Max results to return (1-1000). Omit to fetch all.",
              "type": "integer",
              "exclusiveMinimum": 0,
              "maximum": 1000
            },
            "in": "query",
            "name": "limit",
            "description": "Max results to return (1-1000). Omit to fetch all."
          },
          {
            "schema": {
              "default": 0,
              "description": "Number of results to skip (used with limit)",
              "type": "integer",
              "minimum": 0,
              "maximum": 9007199254740991
            },
            "in": "query",
            "name": "offset",
            "description": "Number of results to skip (used with limit)"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Incidents"
                }
              }
            }
          },
          "304": {
            "description": "Not modified; the ETag the client sent still matches"
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/incidents/count": {
      "get": {
        "operationId": "getIncidentCount",
        "summary": "Count wildlife-vehicle incidents",
        "tags": ["Incidents"],
        "description": "Returns the number of wildlife-vehicle collision incidents matching the same filters as GET /v1/incidents/, without the records.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated list of years (e.g. 2020,2021), required, max 10",
              "example": "2020,2021"
            },
            "in": "query",
            "name": "year",
            "required": true
          },
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated species IDs from GET /v1/incidents/filters",
              "example": "10,18"
            },
            "in": "query",
            "name": "species"
          },
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated service area IDs from GET /v1/incidents/filters",
              "example": "16,17"
            },
            "in": "query",
            "name": "serviceArea"
          },
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated values: MALE, FEMALE, UNKNOWN",
              "example": "MALE,FEMALE,UNKNOWN"
            },
            "in": "query",
            "name": "sex"
          },
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated values: DAY, DAWN, DUSK, DARK, UNKNOWN",
              "example": "DAY,DAWN,DUSK,DARK,UNKNOWN"
            },
            "in": "query",
            "name": "timeOfKill"
          },
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated values: ADULT, YOUNG, UNKNOWN",
              "example": "ADULT,YOUNG,UNKNOWN"
            },
            "in": "query",
            "name": "age"
          },
          {
            "schema": {
              "description": "Start date filter (inclusive, YYYY-MM-DD)",
              "example": "2021-01-01",
              "type": "string",
              "format": "date",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
            },
            "in": "query",
            "name": "startDate",
            "description": "Start date filter (inclusive, YYYY-MM-DD)"
          },
          {
            "schema": {
              "description": "End date filter (inclusive, YYYY-MM-DD)",
              "example": "2021-12-31",
              "type": "string",
              "format": "date",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$"
            },
            "in": "query",
            "name": "endDate",
            "description": "End date filter (inclusive, YYYY-MM-DD)"
          },
          {
            "schema": {
              "type": "string",
              "description": "GeoJSON geometry string for spatial filtering (e.g. Polygon)",
              "example": "{\"type\":\"Polygon\",\"coordinates\":[[[-123.2,49.2],[-123.0,49.2],[-123.0,49.3],[-123.2,49.3],[-123.2,49.2]]]}"
            },
            "in": "query",
            "name": "geometry"
          },
          {
            "schema": {
              "description": "Route corridor filter: start point longitude",
              "type": "number",
              "minimum": -180,
              "maximum": 180
            },
            "in": "query",
            "name": "routeStartLng",
            "description": "Route corridor filter: start point longitude"
          },
          {
            "schema": {
              "description": "Route corridor filter: start point latitude",
              "type": "number",
              "minimum": -90,
              "maximum": 90
            },
            "in": "query",
            "name": "routeStartLat",
            "description": "Route corridor filter: start point latitude"
          },
          {
            "schema": {
              "description": "Route corridor filter: end point longitude",
              "type": "number",
              "minimum": -180,
              "maximum": 180
            },
            "in": "query",
            "name": "routeEndLng",
            "description": "Route corridor filter: end point longitude"
          },
          {
            "schema": {
              "description": "Route corridor filter: end point latitude",
              "type": "number",
              "minimum": -90,
              "maximum": 90
            },
            "in": "query",
            "name": "routeEndLat",
            "description": "Route corridor filter: end point latitude"
          },
          {
            "schema": {
              "default": 250,
              "description": "Route corridor filter: corridor width in metres",
              "type": "number",
              "exclusiveMinimum": 0,
              "maximum": 20000
            },
            "in": "query",
            "name": "routeCorridorM",
            "description": "Route corridor filter: corridor width in metres"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentCount"
                }
              }
            }
          },
          "304": {
            "description": "Not modified; the ETag the client sent still matches"
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/incidents/filters": {
      "get": {
        "operationId": "getIncidentFilters",
        "summary": "Get available incident filter options",
        "tags": ["Incidents"],
        "description": "Returns all available filter values for populating frontend dropdowns.",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentFilters"
                }
              }
            }
          },
          "304": {
            "description": "Not modified; the ETag the client sent still matches"
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "{protocol}://{host}:{port}",
      "description": "Custom Server",
      "variables": {
        "protocol": {
          "enum": ["http", "https"],
          "default": "http",
          "description": "The protocol used to communicate with the server"
        },
        "host": {
          "default": "localhost",
          "description": "The hostname or IP address of the server"
        },
        "port": {
          "default": "3033",
          "description": "The port on which the server is running"
        }
      }
    },
    {
      "url": "http://localhost",
      "description": "Primary Server"
    },
    {
      "url": "http://localhost:3033",
      "description": "Localhost Access (with port)"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}